scripts/requirements.txt
pykrx>=1.0.51
supabase>=2.0.0
python-dotenv>=1.0.0
pandas>=1.5.0
requests>=2.28.0